Fix plot panel auto-complete from incomplete topic
Previously, if you created a plot panel, and e.g. wanted to plot a topic
named `"/drivetrain frc971.control_loops.drivetrain.Output".left_high`
and then typed `output` into the legend, the aforementioned field might
be listed in the auto-complete options, but if you selected it then
it would auto-complete to `"/drivetrain frc971.control_loops.drivetrain.Output".left_high.`.
This was a consequence of the logic checking against this exact scenario
being broken--in particular, at least the JSON and flatbuffer
parsers for MCAP files set `isComplex` to undefined rather than to false.
Given that multiple tests seem to duplicate this assumption, I am
assuming that this is the fault of the MessagePathInput and not the MCAP
parser.